fix files

+++ scripts/reinstall-packages.sh
@@ -1,5 +1,3 @@
-
-
-rm package-lock.json
+if exist package-lock.json rm package-lock.json
 rm -rf node_modules
 npm i -dd
